feat: add wandb param log extractor
